old/requirements.txt
python-dotenv
pydantic
instructor
openai